feat: 서버 에러 핸들링을 위한 커스텀 훅 추가
