+++ src/feature/CardRegister/hooks/useCardForm.ts
@@ -21,12 +21,6 @@ export type CardFormFieldsType = {
   cardCompany: CardCompanyFieldType;
 };
 
-export type ServerFieldErrors = {
-  cardNumber?: string;
-  expirationDate?: string;
-  cvc?: string;
-};
-
 const CARD_FORM_STEP = {
   CARD_NUMBER: 0,
   CARD_COMPANY: 1,
@@ -60,16 +54,6 @@ export const useCardForm = () => {
     onComplete: () => advanceStep(CARD_FORM_STEP.SUBMIT),
   });
 
-  const [serverFieldErrors, setServerFieldErrors] = useState<ServerFieldErrors>(
-    {},
-  );
-  const clearServerFieldError = (fieldName: keyof ServerFieldErrors) => {
-    setServerFieldErrors((prev) => ({
-      ...prev,
-      [fieldName]: undefined,
-    }));
-  };
-
   const fields = {
     numbers: numbersField,
     cardCompany: cardCompanyField,
@@ -101,8 +85,5 @@ export const useCardForm = () => {
     cardFormInfo,
     currentStep,
     hasFormError,
-    serverFieldErrors,
-    setServerFieldErrors,
-    clearServerFieldError,
   };
 };
